Sebez-main/requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
whitenoise==6.2.0
